+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `Matrix4d.setScale(...)` methods to change matrix into a scaling matrix.
 * Add `Matrix4d.scale(...)` methods to support applying a scale operation to a matrix.
 * Change the `Matrix4d.asString()` method to layout the matrix components in multi-line column major format rather than a single line row-major format. Introduce the methods `Matrix4d.asColumnMajorString()` and `Matrix4d.asRowMajorString()` so that the desired string form can be explicitly requested.
+* Add `Matrix4d.setPerspective(...)` method to set matrix up with a perspective projection.
 
 ### [v0.08](https://github.com/realityforge/vecmath/tree/v0.08) (2020-10-24)
 [Full Changelog](https://github.com/realityforge/vecmath/compare/v0.07...v0.08)